Cast time_t to srand seed
Hopefully this'll shut the Windows build's warnings up

(cherry picked from commit 3b11345bfe5b7c8405013a727c853317e8bda013)

+++ src/Utils.h
@@ -8,7 +8,7 @@
 //TODO: Not inline these
 //Call this once, before ever calling one of these `random()` functions
 inline void randomInit() {
-	srand(time(nullptr));
+	srand(static_cast<unsigned int>(time(nullptr)));
 }
 
 inline float random(const float max)
